backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
openai==1.12.0
langchain==0.1.1
faiss-cpu==1.7.4
pydantic==2.5.2
python-dotenv==1.0.0
supabase==2.4.0
numpy==1.26.4
